fix mem leak

+++ dns.h
@@ -20,7 +20,7 @@ typedef struct {
 #define DNS_UPSTREAM_TIMEOUT 500
 #define DNS_UPSTREAM_LISTEN_PORT 12345
 
-extern int32_t transaction_id_counter;
+extern uint16_t transaction_id_counter;
 
 void dns_handle_q(dns_handle_arg_t *arg);
 void dns_handle_r(dns_handle_arg_t *arg);
